refactor: merge index and splash screen into one file
