+++ index.lua
@@ -1,16 +1,63 @@
--- https://sourceforge.net/p/microlua/wiki/API471/
+-- Docs for MicroLua DS: https://sourceforge.net/p/microlua/wiki/API471/
 
--- Load all modules TODO: Load utils here and pass it down to other behaviour
-local splash = dofile("splash.lua")
+local game = require("reel_spin")
+
+-- Load Soundbank and all SFX
 Sound.loadBank("assets/sound/soundbank.bin")
 for i = 0, 4 do
     Sound.loadSFX(i)
 end
 
--- Launch the splash screen
-while not splash() == 1 do
+local function start_splash()
+    -- Always start this sound when entering this screen
+    Sound.startSFX(4)
+
+    local upper_image = Image.load("assets/sprites/Cover.png", VRAM)
+    local lower_image = Image.load("assets/sprites/Cover_lower.png", VRAM)
+
+    while true do
+        Controls.read()
+
+        if Keys.newPress.Down then
+            selected = selected + 1
+            if selected > total_items then selected = 1 end
+        elseif Keys.newPress.Up then
+            selected = selected - 1
+            if selected < 1 then selected = total_items end
+        end
+
+        if Keys.newPress.A then
+            Image.destroy(upper_image)
+            Image.destroy(lower_image)
+
+            Sound.stopAllSFX()
+            game()
+            Sound.startSFX(4)
+
+            upper_image = Image.load("assets/sprites/Cover.png", VRAM)
+            lower_image = Image.load("assets/sprites/Cover_lower.png", VRAM)
+        elseif Keys.newPress.Start then
+            Image.destroy(lower_image)
+            Image.destroy(upper_image)
+            return 1 -- When quitting, we need to return 1 to actually quit cleanly
+        end
+
+        -- Draw Images
+        screen.blit(SCREEN_UP, 0, 0, upper_image)
+        screen.blit(SCREEN_DOWN, 0, 0, lower_image)
+
+        screen.drawFillRect(SCREEN_DOWN, 0, 180, 256, 192, Color.new256(0, 0, 0))
+        screen.print(SCREEN_DOWN, 65, 182, "A: Begin | START: Quit", Color.new256(200, 200, 200))
+
+        render()
+    end
 end
 
+-- Launch the splash screen
+while not start_splash() == 1 do
+end
+
+-- Unload Sounds to clean Memory
 for i = 1, 4 do
     Sound.unloadSFX(i)
 end
